requirements.txt

accelerate==0.18.0
body-visualizer==1.1.0
chumpy==0.70
git+https://github.com/openai/CLIP.git
configer==1.4.1
configparser==5.3.0
diffusers==0.16.0
ftfy==6.1.1
huggingface-hub==0.13.4
human-body-prior==2.2.2.0
icecream==2.1.3
imageio==2.26.0
matplotlib==3.5.3
networkx==2.6.3
neural-renderer-pytorch==1.1.3
numpy==1.21.6
open3d-python==0.7.0.0
opencv-python==4.7.0.72
psbody-mesh==0.4
PyMCubes==0.1.4
pyrender==0.1.45
PyYAML==6.0
smplx==0.1.28
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboardX
tifffile==2021.11.2
tokenizers==0.13.3
torch==1.13.1
torchvision==0.13.1
tqdm==4.65.0
transformers==4.28.1
trimesh==3.9.8
git+https://github.com/facebookresearch/xformers.git
